Changed '==' to '===' in line 6
It is considered good practice to use the type-safe equality operators === and !== instead of their regular counterparts == and !=.

If one of those occurs in an innocent-looking statement such as a == b the actual problem is very difficult to spot.

The strict equality operators (=== and !==) use the strict equality comparison algorithm to compare two operands.

If the operands are of different types, return false.
If both operands are objects, return true only if they refer to the same object.
If both operands are null or both operands are undefined, return true.
If either operand is NaN, return false.
Otherwise, compare the two operand's values:
Numbers must have the same numeric values. +0 and -0 are considered to be the same value.
Strings must have the same characters in the same order.
Booleans must be both true or both false.
The most notable difference between this operator and the equality (==) operator is that if the operands are of different types, the == operator attempts to convert them to the same type before comparing.